pyproject.toml
[project]
name = "ai-emoji"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.6",
    "fastapi>=0.115.12",
    "numpy>=2.2.6",
    "pillow>=11.2.1",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.2",
    "uvicorn>=0.34.3",
]
